+++ examples/docs/ontologies/generated-dir/foaf/entities/Group.ts
@@ -1,6 +1,6 @@
 // ============================================================
 // AUTO-GENERATED — DO NOT EDIT
-// Generated: 2026-05-19T16:53:58.881Z
+// Generated: 2026-05-20T04:39:50.519Z
 // Source:    examples/docs/ontologies/foaf-subset.jsonld
 // IRI:       http://xmlns.com/foaf/0.1/Group
 // ============================================================